ready to publish

+++ lib/ftp_command_handler.dart
@@ -1,9 +1,11 @@
 import 'dart:io';
-import 'package:ftp_server/ftp_session.dart'; // Ensure this path matches your package structure
+import 'package:ftp_server/ftp_session.dart';
+import 'package:ftp_server/server_type.dart';
+// Import the enum
 
 class FTPCommandHandler {
   final Socket controlSocket;
- 
+
   FTPCommandHandler(this.controlSocket);
 
   void handleCommand(String commandLine, FtpSession session) {
@@ -11,20 +13,18 @@ class FTPCommandHandler {
     String command = parts[0].toUpperCase();
     String argument = parts.length > 1 ? parts.sublist(1).join(' ').trim() : '';
     print('Command: $command, Argument: $argument');
-    // Using a switch or if-else chain to handle different commands
+
     switch (command) {
       case 'USER':
-        // Process USER command
         session.cachedUsername = argument;
         session.isAuthenticated =
             false; // Reset authentication status pending password check
         session.sendResponse('331 Password required for $argument');
         break;
       case 'PASS':
-        // Process PASS command, typically you would check against a stored password
         if ((session.username == null && session.password == null) ||
-            (session.cachedUsername == session.username && argument == session.password)) {
-          // Replace with your authentication logic
+            (session.cachedUsername == session.username &&
+                argument == session.password)) {
           session.isAuthenticated = true;
           session.sendResponse('230 User logged in, proceed');
         } else {
@@ -48,23 +48,35 @@ class FTPCommandHandler {
         session.retrieveFile(argument);
         break;
       case 'STOR':
-        session.storeFile(argument);
-        break;
-      case 'PWD':
-        session.sendResponse(
-            '257 "${session.currentDirectory}" is the current directory');
+        if (session.serverType == ServerType.readOnly) {
+          session.sendResponse('550 Command not allowed in read-only mode');
+        } else {
+          session.storeFile(argument);
+        }
         break;
       case 'CWD':
         session.changeDirectory(argument);
         break;
       case 'MKD':
-        session.makeDirectory(argument);
+        if (session.serverType == ServerType.readOnly) {
+          session.sendResponse('550 Command not allowed in read-only mode');
+        } else {
+          session.makeDirectory(argument);
+        }
         break;
       case 'RMD':
-        session.removeDirectory(argument);
+        if (session.serverType == ServerType.readOnly) {
+          session.sendResponse('550 Command not allowed in read-only mode');
+        } else {
+          session.removeDirectory(argument);
+        }
         break;
       case 'DELE':
-        session.deleteFile(argument);
+        if (session.serverType == ServerType.readOnly) {
+          session.sendResponse('550 Command not allowed in read-only mode');
+        } else {
+          session.deleteFile(argument);
+        }
         break;
       case 'SYST':
         session.sendResponse(
@@ -74,7 +86,6 @@ class FTPCommandHandler {
         session.sendResponse('200 NOOP command successful');
         break;
       case 'TYPE':
-        // Typically responds to A (ASCII) or I (binary) types
         if (argument == 'A' || argument == 'I') {
           session.sendResponse('200 Type set to $argument');
         } else {
